refactor: move error message to ui
