+++ packages/ui/package.json
@@ -66,6 +66,7 @@
     "vitest": "catalog:"
   },
   "peerDependencies": {
+    "@junat/react-hooks": "workspace:*",
     "react": "18.x || 19.x",
     "react-dom": "18.x || 19.x"
   },
